requirements.txt
streamlit
langchain
langchain-community
arxiv
pymupdf
sentence-transformers
faiss-cpu
python-dotenv
huggingface-hub